Add files via upload
This yaml file will be used by Ansible. Because we are going to test hundreds of VMs connectivity through our network. So we should automate this process to bring this script to our destination via Ansible. Therefore, some other configuration(Linux commands) will be needed in our destination to make that script executable. This yaml file is the solution instead of logging on each vm one by one and make necessary changes within it.